Updated Gemfile
Would not install on ubuntu server 12.04 for me, had to make changes to gemfile in order for it to connect to github, this is the modified file that worked for me. had to change git:// to https:// not sure why, maybe it is a local issue related to the network we are running here, maybe not, who knows, still someone that knows more than me may be able to work it out?